tests/wpt/web-platform-tests/FileAPI/blob/Blob-close.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Blob.close</title>
<link rel=help href="http://dev.w3.org/2006/webapi/FileAPI/#dfn-close">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/Blob.js"></script>
<div id="log"></div>
<script>
test(function() {
  var blob = new Blob(["TEST"]);
  var sliced = blob.slice();
  blob.close();

  async_test(function(t) {
    var reader = new FileReader();

    reader.onload = t.step_func(function(evt) {
      assert_unreached("Should not dispatch the load event");
    });

    reader.onerror = t.step_func(function(e) {
      assert_equals(reader.result, null);
      assert_equals(reader.error.code, DOMException.INVALID_STATE_ERR);
      t.done();
    });

    reader.readAsText(blob, "UTF-8");
  }, "Closed Blob");

  test_blob(function() {
    return sliced;
  }, {
    expected: "TEST",
    type: "",
    desc: "Slice should still have the data."
  });
});
</script>
